requirements.txt
pandas>=1.3.0
polygon-api-client>=1.0.0
numpy>=1.20.0
